@@ -7,6 +7,8 @@
     "drag and drop",
     "dnd",
     "sortable",
+    "reorder",
+    "reorderable",
     "react",
     "react.js",
     "natural",
@@ -23,23 +25,26 @@
     "/src"
   ],
   "scripts": {
-    "test": "jest --config ./jest.config.js && cross-env SNAPSHOT=check yarn build:dist",
+    "test": "jest --config ./jest.config.js",
     "validate": "yarn run lint && yarn run typecheck",
     "lint": "yarn eslint .",
     "typecheck": "flow check",
+    "check-bundle-sizes": "cross-env SNAPSHOT=check yarn build:dist && yarn build:clean",
+    "update-bundle-sizes": "yarn build:dist && yarn build:clean",
     "build": "yarn run build:clean && yarn run build:dist && yarn run build:flow",
     "build:clean": "rimraf dist",
     "build:dist": "rollup -c",
     "build:flow": "echo \"// @flow\n\nexport * from '../src';\" > dist/react-beautiful-dnd.cjs.js.flow",
     "storybook": "start-storybook -p 9002",
     "build-storybook": "build-storybook -c .storybook -o site",
-    "prepublish": "yarn run build && cd website && yarn",
     "website:start": "cd website && yarn develop",
-    "website:build": "cd website && yarn build"
+    "website:build": "cd website && yarn build",
+    "prepare": "cd website && yarn",
+    "prepublishOnly": "yarn build"
   },
   "dependencies": {
     "babel-runtime": "^6.26.0",
-    "invariant": "^2.2.2",
+    "invariant": "^2.2.4",
     "memoize-one": "^3.1.1",
     "prop-types": "^15.6.0",
     "raf-schd": "^2.1.1",
@@ -64,7 +69,7 @@
     "babel-preset-react": "^6.24.1",
     "cross-env": "^5.1.3",
     "enzyme": "^3.3.0",
-    "enzyme-adapter-react-15": "^1.0.5",
+    "enzyme-adapter-react-16": "^1.1.1",
     "eslint": "^4.15.0",
     "eslint-config-airbnb": "^16.1.0",
     "eslint-plugin-flowtype": "^2.41.0",
@@ -76,9 +81,9 @@
     "jest": "^22.0.5",
     "puppeteer": "^1.2.0",
     "raf-stub": "^2.0.2",
-    "react": "^15.6.2",
-    "react-dom": "^15.6.2",
-    "react-test-renderer": "^16.2.0",
+    "react": "^16.3.1",
+    "react-dom": "^16.3.1",
+    "react-test-renderer": "^16.3.1",
     "rimraf": "^2.6.2",
     "rollup": "^0.56.2",
     "rollup-plugin-babel": "^3.0.3",
@@ -91,7 +96,7 @@
     "styled-components": "^3.2.0"
   },
   "peerDependencies": {
-    "react": ">=15.4.0 || ^16.0.0"
+    "react": "^16.3.1"
   },
   "license": "Apache-2.0",
   "repository": {
