@@ -1,5 +1,5 @@
-// flow-typed signature: 10ddf02cdac165d908159d1ed9162354
-// flow-typed version: <<STUB>>/@atlaskit/css-reset_v1.1.5/flow_v0.59.0
+// flow-typed signature: f7743c067ed24cbefc7e7b3286211ab5
+// flow-typed version: <<STUB>>/@atlaskit/css-reset_v^1.2.0/flow_v0.68.0
 
 /**
  * This is an autogenerated libdef stub for:
