@@ -1,5 +1,5 @@
-// flow-typed signature: 2315204a2a7137e2106809890166e4f3
-// flow-typed version: <<STUB>>/@storybook/react_v3.2.16/flow_v0.59.0
+// flow-typed signature: e1a3dbf32dba20faf93e94de9829f082
+// flow-typed version: <<STUB>>/@storybook/react_v^3.2.16/flow_v0.68.0
 
 /**
  * This is an autogenerated libdef stub for:
@@ -154,14 +154,6 @@ declare module '@storybook/react/dist/server/config/webpack.config.prod' {
   declare module.exports: any;
 }
 
-declare module '@storybook/react/dist/server/iframe.html' {
-  declare module.exports: any;
-}
-
-declare module '@storybook/react/dist/server/index.html' {
-  declare module.exports: any;
-}
-
 declare module '@storybook/react/dist/server/index' {
   declare module.exports: any;
 }
@@ -171,6 +163,150 @@ declare module '@storybook/react/dist/server/middleware' {
 }
 
 declare module '@storybook/react/dist/server/utils' {
+  declare module.exports: any;
+}
+
+declare module '@storybook/react/src/client/index' {
+  declare module.exports: any;
+}
+
+declare module '@storybook/react/src/client/manager/index' {
+  declare module.exports: any;
+}
+
+declare module '@storybook/react/src/client/manager/preview' {
+  declare module.exports: any;
+}
+
+declare module '@storybook/react/src/client/manager/provider' {
+  declare module.exports: any;
+}
+
+declare module '@storybook/react/src/client/preview/actions' {
+  declare module.exports: any;
+}
+
+declare module '@storybook/react/src/client/preview/client_api' {
+  declare module.exports: any;
+}
+
+declare module '@storybook/react/src/client/preview/client_api.test' {
+  declare module.exports: any;
+}
+
+declare module '@storybook/react/src/client/preview/config_api' {
+  declare module.exports: any;
+}
+
+declare module '@storybook/react/src/client/preview/element_check' {
+  declare module.exports: any;
+}
+
+declare module '@storybook/react/src/client/preview/element_check.test' {
+  declare module.exports: any;
+}
+
+declare module '@storybook/react/src/client/preview/error_display' {
+  declare module.exports: any;
+}
+
+declare module '@storybook/react/src/client/preview/index' {
+  declare module.exports: any;
+}
+
+declare module '@storybook/react/src/client/preview/init' {
+  declare module.exports: any;
+}
+
+declare module '@storybook/react/src/client/preview/reducer' {
+  declare module.exports: any;
+}
+
+declare module '@storybook/react/src/client/preview/render' {
+  declare module.exports: any;
+}
+
+declare module '@storybook/react/src/client/preview/story_store' {
+  declare module.exports: any;
+}
+
+declare module '@storybook/react/src/demo/Button' {
+  declare module.exports: any;
+}
+
+declare module '@storybook/react/src/demo/Welcome' {
+  declare module.exports: any;
+}
+
+declare module '@storybook/react/src/server/addons' {
+  declare module.exports: any;
+}
+
+declare module '@storybook/react/src/server/babel_config' {
+  declare module.exports: any;
+}
+
+declare module '@storybook/react/src/server/babel_config.test' {
+  declare module.exports: any;
+}
+
+declare module '@storybook/react/src/server/build' {
+  declare module.exports: any;
+}
+
+declare module '@storybook/react/src/server/config' {
+  declare module.exports: any;
+}
+
+declare module '@storybook/react/src/server/config/babel' {
+  declare module.exports: any;
+}
+
+declare module '@storybook/react/src/server/config/babel.prod' {
+  declare module.exports: any;
+}
+
+declare module '@storybook/react/src/server/config/defaults/webpack.config' {
+  declare module.exports: any;
+}
+
+declare module '@storybook/react/src/server/config/globals' {
+  declare module.exports: any;
+}
+
+declare module '@storybook/react/src/server/config/polyfills' {
+  declare module.exports: any;
+}
+
+declare module '@storybook/react/src/server/config/utils' {
+  declare module.exports: any;
+}
+
+declare module '@storybook/react/src/server/config/WatchMissingNodeModulesPlugin' {
+  declare module.exports: any;
+}
+
+declare module '@storybook/react/src/server/config/webpack.config' {
+  declare module.exports: any;
+}
+
+declare module '@storybook/react/src/server/config/webpack.config.prod' {
+  declare module.exports: any;
+}
+
+declare module '@storybook/react/src/server/index' {
+  declare module.exports: any;
+}
+
+declare module '@storybook/react/src/server/middleware' {
+  declare module.exports: any;
+}
+
+declare module '@storybook/react/src/server/utils' {
+  declare module.exports: any;
+}
+
+declare module '@storybook/react/src/server/utils.test' {
   declare module.exports: any;
 }
 
@@ -274,12 +410,6 @@ declare module '@storybook/react/dist/server/config/webpack.config.js' {
 declare module '@storybook/react/dist/server/config/webpack.config.prod.js' {
   declare module.exports: $Exports<'@storybook/react/dist/server/config/webpack.config.prod'>;
 }
-declare module '@storybook/react/dist/server/iframe.html.js' {
-  declare module.exports: $Exports<'@storybook/react/dist/server/iframe.html'>;
-}
-declare module '@storybook/react/dist/server/index.html.js' {
-  declare module.exports: $Exports<'@storybook/react/dist/server/index.html'>;
-}
 declare module '@storybook/react/dist/server/index.js' {
   declare module.exports: $Exports<'@storybook/react/dist/server/index'>;
 }
@@ -288,4 +418,112 @@ declare module '@storybook/react/dist/server/middleware.js' {
 }
 declare module '@storybook/react/dist/server/utils.js' {
   declare module.exports: $Exports<'@storybook/react/dist/server/utils'>;
+}
+declare module '@storybook/react/src/client/index.js' {
+  declare module.exports: $Exports<'@storybook/react/src/client/index'>;
+}
+declare module '@storybook/react/src/client/manager/index.js' {
+  declare module.exports: $Exports<'@storybook/react/src/client/manager/index'>;
+}
+declare module '@storybook/react/src/client/manager/preview.js' {
+  declare module.exports: $Exports<'@storybook/react/src/client/manager/preview'>;
+}
+declare module '@storybook/react/src/client/manager/provider.js' {
+  declare module.exports: $Exports<'@storybook/react/src/client/manager/provider'>;
+}
+declare module '@storybook/react/src/client/preview/actions.js' {
+  declare module.exports: $Exports<'@storybook/react/src/client/preview/actions'>;
+}
+declare module '@storybook/react/src/client/preview/client_api.js' {
+  declare module.exports: $Exports<'@storybook/react/src/client/preview/client_api'>;
+}
+declare module '@storybook/react/src/client/preview/client_api.test.js' {
+  declare module.exports: $Exports<'@storybook/react/src/client/preview/client_api.test'>;
+}
+declare module '@storybook/react/src/client/preview/config_api.js' {
+  declare module.exports: $Exports<'@storybook/react/src/client/preview/config_api'>;
+}
+declare module '@storybook/react/src/client/preview/element_check.js' {
+  declare module.exports: $Exports<'@storybook/react/src/client/preview/element_check'>;
+}
+declare module '@storybook/react/src/client/preview/element_check.test.js' {
+  declare module.exports: $Exports<'@storybook/react/src/client/preview/element_check.test'>;
+}
+declare module '@storybook/react/src/client/preview/error_display.js' {
+  declare module.exports: $Exports<'@storybook/react/src/client/preview/error_display'>;
+}
+declare module '@storybook/react/src/client/preview/index.js' {
+  declare module.exports: $Exports<'@storybook/react/src/client/preview/index'>;
+}
+declare module '@storybook/react/src/client/preview/init.js' {
+  declare module.exports: $Exports<'@storybook/react/src/client/preview/init'>;
+}
+declare module '@storybook/react/src/client/preview/reducer.js' {
+  declare module.exports: $Exports<'@storybook/react/src/client/preview/reducer'>;
+}
+declare module '@storybook/react/src/client/preview/render.js' {
+  declare module.exports: $Exports<'@storybook/react/src/client/preview/render'>;
+}
+declare module '@storybook/react/src/client/preview/story_store.js' {
+  declare module.exports: $Exports<'@storybook/react/src/client/preview/story_store'>;
+}
+declare module '@storybook/react/src/demo/Button.js' {
+  declare module.exports: $Exports<'@storybook/react/src/demo/Button'>;
+}
+declare module '@storybook/react/src/demo/Welcome.js' {
+  declare module.exports: $Exports<'@storybook/react/src/demo/Welcome'>;
+}
+declare module '@storybook/react/src/server/addons.js' {
+  declare module.exports: $Exports<'@storybook/react/src/server/addons'>;
+}
+declare module '@storybook/react/src/server/babel_config.js' {
+  declare module.exports: $Exports<'@storybook/react/src/server/babel_config'>;
+}
+declare module '@storybook/react/src/server/babel_config.test.js' {
+  declare module.exports: $Exports<'@storybook/react/src/server/babel_config.test'>;
+}
+declare module '@storybook/react/src/server/build.js' {
+  declare module.exports: $Exports<'@storybook/react/src/server/build'>;
+}
+declare module '@storybook/react/src/server/config.js' {
+  declare module.exports: $Exports<'@storybook/react/src/server/config'>;
+}
+declare module '@storybook/react/src/server/config/babel.js' {
+  declare module.exports: $Exports<'@storybook/react/src/server/config/babel'>;
+}
+declare module '@storybook/react/src/server/config/babel.prod.js' {
+  declare module.exports: $Exports<'@storybook/react/src/server/config/babel.prod'>;
+}
+declare module '@storybook/react/src/server/config/defaults/webpack.config.js' {
+  declare module.exports: $Exports<'@storybook/react/src/server/config/defaults/webpack.config'>;
+}
+declare module '@storybook/react/src/server/config/globals.js' {
+  declare module.exports: $Exports<'@storybook/react/src/server/config/globals'>;
+}
+declare module '@storybook/react/src/server/config/polyfills.js' {
+  declare module.exports: $Exports<'@storybook/react/src/server/config/polyfills'>;
+}
+declare module '@storybook/react/src/server/config/utils.js' {
+  declare module.exports: $Exports<'@storybook/react/src/server/config/utils'>;
+}
+declare module '@storybook/react/src/server/config/WatchMissingNodeModulesPlugin.js' {
+  declare module.exports: $Exports<'@storybook/react/src/server/config/WatchMissingNodeModulesPlugin'>;
+}
+declare module '@storybook/react/src/server/config/webpack.config.js' {
+  declare module.exports: $Exports<'@storybook/react/src/server/config/webpack.config'>;
+}
+declare module '@storybook/react/src/server/config/webpack.config.prod.js' {
+  declare module.exports: $Exports<'@storybook/react/src/server/config/webpack.config.prod'>;
+}
+declare module '@storybook/react/src/server/index.js' {
+  declare module.exports: $Exports<'@storybook/react/src/server/index'>;
+}
+declare module '@storybook/react/src/server/middleware.js' {
+  declare module.exports: $Exports<'@storybook/react/src/server/middleware'>;
+}
+declare module '@storybook/react/src/server/utils.js' {
+  declare module.exports: $Exports<'@storybook/react/src/server/utils'>;
+}
+declare module '@storybook/react/src/server/utils.test.js' {
+  declare module.exports: $Exports<'@storybook/react/src/server/utils.test'>;
 }
