@@ -1,5 +1,5 @@
-// flow-typed signature: e5782a4b28ce2bc8708cae5117dc642b
-// flow-typed version: <<STUB>>/babel-cli_v6.26.0/flow_v0.59.0
+// flow-typed signature: f378965f8d54d704ec94462b6a297598
+// flow-typed version: <<STUB>>/babel-cli_v^6.26.0/flow_v0.68.0
 
 /**
  * This is an autogenerated libdef stub for:
