@@ -1,5 +1,5 @@
-// flow-typed signature: 038df3f0f335da51be96fe734211f36a
-// flow-typed version: <<STUB>>/babel-eslint_v8.0.2/flow_v0.59.0
+// flow-typed signature: a3d1544ce880d0d829224209752ea0cd
+// flow-typed version: <<STUB>>/babel-eslint_v^8.2.1/flow_v0.68.0
 
 /**
  * This is an autogenerated libdef stub for:
@@ -22,59 +22,102 @@ declare module 'babel-eslint' {
  * require those files directly. Feel free to delete any files that aren't
  * needed.
  */
-declare module 'babel-eslint/babylon-to-espree/attachComments' {
+declare module 'babel-eslint/lib/analyze-scope' {
   declare module.exports: any;
 }
 
-declare module 'babel-eslint/babylon-to-espree/convertComments' {
+declare module 'babel-eslint/lib/babylon-to-espree/attachComments' {
   declare module.exports: any;
 }
 
-declare module 'babel-eslint/babylon-to-espree/convertTemplateType' {
+declare module 'babel-eslint/lib/babylon-to-espree/convertComments' {
   declare module.exports: any;
 }
 
-declare module 'babel-eslint/babylon-to-espree/index' {
+declare module 'babel-eslint/lib/babylon-to-espree/convertTemplateType' {
   declare module.exports: any;
 }
 
-declare module 'babel-eslint/babylon-to-espree/toAST' {
+declare module 'babel-eslint/lib/babylon-to-espree/index' {
   declare module.exports: any;
 }
 
-declare module 'babel-eslint/babylon-to-espree/toToken' {
+declare module 'babel-eslint/lib/babylon-to-espree/toAST' {
   declare module.exports: any;
 }
 
-declare module 'babel-eslint/babylon-to-espree/toTokens' {
+declare module 'babel-eslint/lib/babylon-to-espree/toToken' {
+  declare module.exports: any;
+}
+
+declare module 'babel-eslint/lib/babylon-to-espree/toTokens' {
+  declare module.exports: any;
+}
+
+declare module 'babel-eslint/lib/index' {
+  declare module.exports: any;
+}
+
+declare module 'babel-eslint/lib/parse-with-patch' {
+  declare module.exports: any;
+}
+
+declare module 'babel-eslint/lib/parse-with-scope' {
+  declare module.exports: any;
+}
+
+declare module 'babel-eslint/lib/parse' {
+  declare module.exports: any;
+}
+
+declare module 'babel-eslint/lib/patch-eslint-scope' {
+  declare module.exports: any;
+}
+
+declare module 'babel-eslint/lib/visitor-keys' {
   declare module.exports: any;
 }
 
 // Filename aliases
-declare module 'babel-eslint/babylon-to-espree/attachComments.js' {
-  declare module.exports: $Exports<'babel-eslint/babylon-to-espree/attachComments'>;
+declare module 'babel-eslint/lib/analyze-scope.js' {
+  declare module.exports: $Exports<'babel-eslint/lib/analyze-scope'>;
 }
-declare module 'babel-eslint/babylon-to-espree/convertComments.js' {
-  declare module.exports: $Exports<'babel-eslint/babylon-to-espree/convertComments'>;
+declare module 'babel-eslint/lib/babylon-to-espree/attachComments.js' {
+  declare module.exports: $Exports<'babel-eslint/lib/babylon-to-espree/attachComments'>;
 }
-declare module 'babel-eslint/babylon-to-espree/convertTemplateType.js' {
-  declare module.exports: $Exports<'babel-eslint/babylon-to-espree/convertTemplateType'>;
+declare module 'babel-eslint/lib/babylon-to-espree/convertComments.js' {
+  declare module.exports: $Exports<'babel-eslint/lib/babylon-to-espree/convertComments'>;
 }
-declare module 'babel-eslint/babylon-to-espree/index.js' {
-  declare module.exports: $Exports<'babel-eslint/babylon-to-espree/index'>;
+declare module 'babel-eslint/lib/babylon-to-espree/convertTemplateType.js' {
+  declare module.exports: $Exports<'babel-eslint/lib/babylon-to-espree/convertTemplateType'>;
 }
-declare module 'babel-eslint/babylon-to-espree/toAST.js' {
-  declare module.exports: $Exports<'babel-eslint/babylon-to-espree/toAST'>;
+declare module 'babel-eslint/lib/babylon-to-espree/index.js' {
+  declare module.exports: $Exports<'babel-eslint/lib/babylon-to-espree/index'>;
 }
-declare module 'babel-eslint/babylon-to-espree/toToken.js' {
-  declare module.exports: $Exports<'babel-eslint/babylon-to-espree/toToken'>;
+declare module 'babel-eslint/lib/babylon-to-espree/toAST.js' {
+  declare module.exports: $Exports<'babel-eslint/lib/babylon-to-espree/toAST'>;
 }
-declare module 'babel-eslint/babylon-to-espree/toTokens.js' {
-  declare module.exports: $Exports<'babel-eslint/babylon-to-espree/toTokens'>;
+declare module 'babel-eslint/lib/babylon-to-espree/toToken.js' {
+  declare module.exports: $Exports<'babel-eslint/lib/babylon-to-espree/toToken'>;
 }
-declare module 'babel-eslint/index' {
-  declare module.exports: $Exports<'babel-eslint'>;
+declare module 'babel-eslint/lib/babylon-to-espree/toTokens.js' {
+  declare module.exports: $Exports<'babel-eslint/lib/babylon-to-espree/toTokens'>;
 }
-declare module 'babel-eslint/index.js' {
-  declare module.exports: $Exports<'babel-eslint'>;
+declare module 'babel-eslint/lib/index.js' {
+  declare module.exports: $Exports<'babel-eslint/lib/index'>;
+}
+declare module 'babel-eslint/lib/parse-with-patch.js' {
+  declare module.exports: $Exports<'babel-eslint/lib/parse-with-patch'>;
+}
+declare module 'babel-eslint/lib/parse-with-scope.js' {
+  declare module.exports: $Exports<'babel-eslint/lib/parse-with-scope'>;
+}
+declare module 'babel-eslint/lib/parse.js' {
+  declare module.exports: $Exports<'babel-eslint/lib/parse'>;
+}
+declare module 'babel-eslint/lib/patch-eslint-scope.js' {
+  declare module.exports: $Exports<'babel-eslint/lib/patch-eslint-scope'>;
+}
+declare module 'babel-eslint/lib/visitor-keys.js' {
+  declare module.exports: $Exports<'babel-eslint/lib/visitor-keys'>;
 }
