@@ -1,5 +1,5 @@
-// flow-typed signature: 45b5082352727867da839e8477a13595
-// flow-typed version: <<STUB>>/babel-plugin-transform-class-properties_v6.24.1/flow_v0.59.0
+// flow-typed signature: 9ff098d5f5575bc71487081f66ba0289
+// flow-typed version: <<STUB>>/babel-plugin-transform-class-properties_v^6.24.1/flow_v0.68.0
 
 /**
  * This is an autogenerated libdef stub for:
