@@ -1,5 +1,5 @@
-// flow-typed signature: 912e1b3ad0e25992485335000f8d1360
-// flow-typed version: <<STUB>>/babel-plugin-transform-export-extensions_v6.22.0/flow_v0.59.0
+// flow-typed signature: 031c0cc3657a8ede06bbac658edfeedf
+// flow-typed version: <<STUB>>/babel-plugin-transform-export-extensions_v^6.22.0/flow_v0.68.0
 
 /**
  * This is an autogenerated libdef stub for:
