@@ -1,5 +1,5 @@
-// flow-typed signature: 804ac1e25ae65003b2023b225fe9b4ce
-// flow-typed version: <<STUB>>/babel-plugin-transform-object-rest-spread_v6.26.0/flow_v0.59.0
+// flow-typed signature: a87bd5d283c6b90f2b0550720106a228
+// flow-typed version: <<STUB>>/babel-plugin-transform-object-rest-spread_v^6.26.0/flow_v0.68.0
 
 /**
  * This is an autogenerated libdef stub for:
