@@ -1,5 +1,5 @@
-// flow-typed signature: 837e437aa4ac506085e7c17b222beb3b
-// flow-typed version: <<STUB>>/babel-plugin-transform-runtime_v6.23.0/flow_v0.59.0
+// flow-typed signature: b666b687bc2b3d8f0a00279ae8cc7224
+// flow-typed version: <<STUB>>/babel-plugin-transform-runtime_v^6.23.0/flow_v0.68.0
 
 /**
  * This is an autogenerated libdef stub for:
