@@ -1,5 +1,5 @@
-// flow-typed signature: 0e8c5deb4891a489cef731359da9db99
-// flow-typed version: <<STUB>>/babel-preset-flow_v6.23.0/flow_v0.59.0
+// flow-typed signature: 51675ed83c30801e379a24c02e22490d
+// flow-typed version: <<STUB>>/babel-preset-flow_v^6.23.0/flow_v0.68.0
 
 /**
  * This is an autogenerated libdef stub for:
