@@ -1,5 +1,5 @@
-// flow-typed signature: 9b8927fa8bd987e0f6eaf5985eb2f153
-// flow-typed version: <<STUB>>/babel-preset-react_v6.24.1/flow_v0.59.0
+// flow-typed signature: 25725911f542289385276583596739d1
+// flow-typed version: <<STUB>>/babel-preset-react_v^6.24.1/flow_v0.68.0
 
 /**
  * This is an autogenerated libdef stub for:
