@@ -1,5 +1,5 @@
-// flow-typed signature: 3dc8a780063a3dba26202f10ccf15c08
-// flow-typed version: <<STUB>>/babel-runtime_v6.26.0/flow_v0.59.0
+// flow-typed signature: 04bb0a26417f872bf56ab01c17a9b3ba
+// flow-typed version: <<STUB>>/babel-runtime_v^6.26.0/flow_v0.68.0
 
 /**
  * This is an autogenerated libdef stub for:
