@@ -1,5 +1,5 @@
-// flow-typed signature: 270bfcf00a2c84306b61c0f8ab5b1c31
-// flow-typed version: <<STUB>>/eslint-config-airbnb_v16.1.0/flow_v0.59.0
+// flow-typed signature: 72ba8dfd359cea3b8c1efd96827fc118
+// flow-typed version: <<STUB>>/eslint-config-airbnb_v^16.1.0/flow_v0.68.0
 
 /**
  * This is an autogenerated libdef stub for:
