@@ -1,5 +1,5 @@
-// flow-typed signature: 55b6e43aed3db209881ee02314c55bbc
-// flow-typed version: <<STUB>>/eslint-plugin-import_v2.8.0/flow_v0.59.0
+// flow-typed signature: 9ae35733d7ccd3dd4f4ae010499e7aa9
+// flow-typed version: <<STUB>>/eslint-plugin-import_v^2.8.0/flow_v0.68.0
 
 /**
  * This is an autogenerated libdef stub for:
@@ -58,6 +58,10 @@ declare module 'eslint-plugin-import/lib/core/staticRequire' {
   declare module.exports: any;
 }
 
+declare module 'eslint-plugin-import/lib/docsUrl' {
+  declare module.exports: any;
+}
+
 declare module 'eslint-plugin-import/lib/ExportMap' {
   declare module.exports: any;
 }
@@ -87,6 +91,10 @@ declare module 'eslint-plugin-import/lib/rules/extensions' {
 }
 
 declare module 'eslint-plugin-import/lib/rules/first' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-import/lib/rules/group-exports' {
   declare module.exports: any;
 }
 
@@ -123,6 +131,10 @@ declare module 'eslint-plugin-import/lib/rules/no-anonymous-default-export' {
 }
 
 declare module 'eslint-plugin-import/lib/rules/no-commonjs' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-import/lib/rules/no-default-export' {
   declare module.exports: any;
 }
 
@@ -174,11 +186,19 @@ declare module 'eslint-plugin-import/lib/rules/no-restricted-paths' {
   declare module.exports: any;
 }
 
+declare module 'eslint-plugin-import/lib/rules/no-self-import' {
+  declare module.exports: any;
+}
+
 declare module 'eslint-plugin-import/lib/rules/no-unassigned-import' {
   declare module.exports: any;
 }
 
 declare module 'eslint-plugin-import/lib/rules/no-unresolved' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-import/lib/rules/no-useless-path-segments' {
   declare module.exports: any;
 }
 
@@ -230,6 +250,9 @@ declare module 'eslint-plugin-import/lib/core/importType.js' {
 declare module 'eslint-plugin-import/lib/core/staticRequire.js' {
   declare module.exports: $Exports<'eslint-plugin-import/lib/core/staticRequire'>;
 }
+declare module 'eslint-plugin-import/lib/docsUrl.js' {
+  declare module.exports: $Exports<'eslint-plugin-import/lib/docsUrl'>;
+}
 declare module 'eslint-plugin-import/lib/ExportMap.js' {
   declare module.exports: $Exports<'eslint-plugin-import/lib/ExportMap'>;
 }
@@ -253,6 +276,9 @@ declare module 'eslint-plugin-import/lib/rules/extensions.js' {
 }
 declare module 'eslint-plugin-import/lib/rules/first.js' {
   declare module.exports: $Exports<'eslint-plugin-import/lib/rules/first'>;
+}
+declare module 'eslint-plugin-import/lib/rules/group-exports.js' {
+  declare module.exports: $Exports<'eslint-plugin-import/lib/rules/group-exports'>;
 }
 declare module 'eslint-plugin-import/lib/rules/imports-first.js' {
   declare module.exports: $Exports<'eslint-plugin-import/lib/rules/imports-first'>;
@@ -280,6 +306,9 @@ declare module 'eslint-plugin-import/lib/rules/no-anonymous-default-export.js' {
 }
 declare module 'eslint-plugin-import/lib/rules/no-commonjs.js' {
   declare module.exports: $Exports<'eslint-plugin-import/lib/rules/no-commonjs'>;
+}
+declare module 'eslint-plugin-import/lib/rules/no-default-export.js' {
+  declare module.exports: $Exports<'eslint-plugin-import/lib/rules/no-default-export'>;
 }
 declare module 'eslint-plugin-import/lib/rules/no-deprecated.js' {
   declare module.exports: $Exports<'eslint-plugin-import/lib/rules/no-deprecated'>;
@@ -317,11 +346,17 @@ declare module 'eslint-plugin-import/lib/rules/no-nodejs-modules.js' {
 declare module 'eslint-plugin-import/lib/rules/no-restricted-paths.js' {
   declare module.exports: $Exports<'eslint-plugin-import/lib/rules/no-restricted-paths'>;
 }
+declare module 'eslint-plugin-import/lib/rules/no-self-import.js' {
+  declare module.exports: $Exports<'eslint-plugin-import/lib/rules/no-self-import'>;
+}
 declare module 'eslint-plugin-import/lib/rules/no-unassigned-import.js' {
   declare module.exports: $Exports<'eslint-plugin-import/lib/rules/no-unassigned-import'>;
 }
 declare module 'eslint-plugin-import/lib/rules/no-unresolved.js' {
   declare module.exports: $Exports<'eslint-plugin-import/lib/rules/no-unresolved'>;
+}
+declare module 'eslint-plugin-import/lib/rules/no-useless-path-segments.js' {
+  declare module.exports: $Exports<'eslint-plugin-import/lib/rules/no-useless-path-segments'>;
 }
 declare module 'eslint-plugin-import/lib/rules/no-webpack-loader-syntax.js' {
   declare module.exports: $Exports<'eslint-plugin-import/lib/rules/no-webpack-loader-syntax'>;
