@@ -1,5 +1,5 @@
-// flow-typed signature: 558356434ef2050af5f029eedd0d69b6
-// flow-typed version: <<STUB>>/eslint-plugin-jest_v21.3.2/flow_v0.59.0
+// flow-typed signature: 9fe756f352d699ac7acf77bbcec31bb8
+// flow-typed version: <<STUB>>/eslint-plugin-jest_v^21.6.1/flow_v0.68.0
 
 /**
  * This is an autogenerated libdef stub for:
@@ -22,43 +22,143 @@ declare module 'eslint-plugin-jest' {
  * require those files directly. Feel free to delete any files that aren't
  * needed.
  */
-declare module 'eslint-plugin-jest/rules/__tests__/no_focused_tests.test' {
+declare module 'eslint-plugin-jest/processors/__tests__/snapshot-processor.test' {
   declare module.exports: any;
 }
 
-declare module 'eslint-plugin-jest/rules/__tests__/no_identical_title.test' {
+declare module 'eslint-plugin-jest/processors/snapshot-processor' {
   declare module.exports: any;
 }
 
-declare module 'eslint-plugin-jest/rules/__tests__/no_skipped_tests.test' {
+declare module 'eslint-plugin-jest/rules/__tests__/consistent-test-it.test' {
   declare module.exports: any;
 }
 
-declare module 'eslint-plugin-jest/rules/__tests__/prefer_to_have_length.test' {
+declare module 'eslint-plugin-jest/rules/__tests__/lowercase-name.test' {
   declare module.exports: any;
 }
 
-declare module 'eslint-plugin-jest/rules/__tests__/valid_expect.test' {
+declare module 'eslint-plugin-jest/rules/__tests__/no-disabled-tests.test' {
   declare module.exports: any;
 }
 
-declare module 'eslint-plugin-jest/rules/no_disabled_tests' {
+declare module 'eslint-plugin-jest/rules/__tests__/no-focused-tests.test' {
   declare module.exports: any;
 }
 
-declare module 'eslint-plugin-jest/rules/no_focused_tests' {
+declare module 'eslint-plugin-jest/rules/__tests__/no-hooks.test' {
   declare module.exports: any;
 }
 
-declare module 'eslint-plugin-jest/rules/no_identical_title' {
+declare module 'eslint-plugin-jest/rules/__tests__/no-identical-title.test' {
   declare module.exports: any;
 }
 
-declare module 'eslint-plugin-jest/rules/prefer_to_have_length' {
+declare module 'eslint-plugin-jest/rules/__tests__/no-jest-import.test' {
   declare module.exports: any;
 }
 
-declare module 'eslint-plugin-jest/rules/valid_expect' {
+declare module 'eslint-plugin-jest/rules/__tests__/no-large-snapshots.test' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jest/rules/__tests__/no-test-prefixes.test' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jest/rules/__tests__/prefer-expect-assertions.test' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jest/rules/__tests__/prefer-to-be-null.test' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jest/rules/__tests__/prefer-to-be-undefined.test' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jest/rules/__tests__/prefer-to-have-length.test' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jest/rules/__tests__/valid-describe.test' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jest/rules/__tests__/valid-expect-in-promise.test' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jest/rules/__tests__/valid-expect.test' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jest/rules/consistent-test-it' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jest/rules/lowercase-name' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jest/rules/no-disabled-tests' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jest/rules/no-focused-tests' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jest/rules/no-hooks' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jest/rules/no-identical-title' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jest/rules/no-jest-import' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jest/rules/no-large-snapshots' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jest/rules/no-test-prefixes' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jest/rules/prefer-expect-assertions' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jest/rules/prefer-to-be-null' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jest/rules/prefer-to-be-undefined' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jest/rules/prefer-to-have-length' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jest/rules/util' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jest/rules/valid-describe' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jest/rules/valid-expect-in-promise' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jest/rules/valid-expect' {
   declare module.exports: any;
 }
 
@@ -69,33 +169,108 @@ declare module 'eslint-plugin-jest/index' {
 declare module 'eslint-plugin-jest/index.js' {
   declare module.exports: $Exports<'eslint-plugin-jest'>;
 }
-declare module 'eslint-plugin-jest/rules/__tests__/no_focused_tests.test.js' {
-  declare module.exports: $Exports<'eslint-plugin-jest/rules/__tests__/no_focused_tests.test'>;
+declare module 'eslint-plugin-jest/processors/__tests__/snapshot-processor.test.js' {
+  declare module.exports: $Exports<'eslint-plugin-jest/processors/__tests__/snapshot-processor.test'>;
 }
-declare module 'eslint-plugin-jest/rules/__tests__/no_identical_title.test.js' {
-  declare module.exports: $Exports<'eslint-plugin-jest/rules/__tests__/no_identical_title.test'>;
+declare module 'eslint-plugin-jest/processors/snapshot-processor.js' {
+  declare module.exports: $Exports<'eslint-plugin-jest/processors/snapshot-processor'>;
 }
-declare module 'eslint-plugin-jest/rules/__tests__/no_skipped_tests.test.js' {
-  declare module.exports: $Exports<'eslint-plugin-jest/rules/__tests__/no_skipped_tests.test'>;
+declare module 'eslint-plugin-jest/rules/__tests__/consistent-test-it.test.js' {
+  declare module.exports: $Exports<'eslint-plugin-jest/rules/__tests__/consistent-test-it.test'>;
 }
-declare module 'eslint-plugin-jest/rules/__tests__/prefer_to_have_length.test.js' {
-  declare module.exports: $Exports<'eslint-plugin-jest/rules/__tests__/prefer_to_have_length.test'>;
+declare module 'eslint-plugin-jest/rules/__tests__/lowercase-name.test.js' {
+  declare module.exports: $Exports<'eslint-plugin-jest/rules/__tests__/lowercase-name.test'>;
 }
-declare module 'eslint-plugin-jest/rules/__tests__/valid_expect.test.js' {
-  declare module.exports: $Exports<'eslint-plugin-jest/rules/__tests__/valid_expect.test'>;
+declare module 'eslint-plugin-jest/rules/__tests__/no-disabled-tests.test.js' {
+  declare module.exports: $Exports<'eslint-plugin-jest/rules/__tests__/no-disabled-tests.test'>;
 }
-declare module 'eslint-plugin-jest/rules/no_disabled_tests.js' {
-  declare module.exports: $Exports<'eslint-plugin-jest/rules/no_disabled_tests'>;
+declare module 'eslint-plugin-jest/rules/__tests__/no-focused-tests.test.js' {
+  declare module.exports: $Exports<'eslint-plugin-jest/rules/__tests__/no-focused-tests.test'>;
 }
-declare module 'eslint-plugin-jest/rules/no_focused_tests.js' {
-  declare module.exports: $Exports<'eslint-plugin-jest/rules/no_focused_tests'>;
+declare module 'eslint-plugin-jest/rules/__tests__/no-hooks.test.js' {
+  declare module.exports: $Exports<'eslint-plugin-jest/rules/__tests__/no-hooks.test'>;
 }
-declare module 'eslint-plugin-jest/rules/no_identical_title.js' {
-  declare module.exports: $Exports<'eslint-plugin-jest/rules/no_identical_title'>;
+declare module 'eslint-plugin-jest/rules/__tests__/no-identical-title.test.js' {
+  declare module.exports: $Exports<'eslint-plugin-jest/rules/__tests__/no-identical-title.test'>;
 }
-declare module 'eslint-plugin-jest/rules/prefer_to_have_length.js' {
-  declare module.exports: $Exports<'eslint-plugin-jest/rules/prefer_to_have_length'>;
+declare module 'eslint-plugin-jest/rules/__tests__/no-jest-import.test.js' {
+  declare module.exports: $Exports<'eslint-plugin-jest/rules/__tests__/no-jest-import.test'>;
 }
-declare module 'eslint-plugin-jest/rules/valid_expect.js' {
-  declare module.exports: $Exports<'eslint-plugin-jest/rules/valid_expect'>;
+declare module 'eslint-plugin-jest/rules/__tests__/no-large-snapshots.test.js' {
+  declare module.exports: $Exports<'eslint-plugin-jest/rules/__tests__/no-large-snapshots.test'>;
+}
+declare module 'eslint-plugin-jest/rules/__tests__/no-test-prefixes.test.js' {
+  declare module.exports: $Exports<'eslint-plugin-jest/rules/__tests__/no-test-prefixes.test'>;
+}
+declare module 'eslint-plugin-jest/rules/__tests__/prefer-expect-assertions.test.js' {
+  declare module.exports: $Exports<'eslint-plugin-jest/rules/__tests__/prefer-expect-assertions.test'>;
+}
+declare module 'eslint-plugin-jest/rules/__tests__/prefer-to-be-null.test.js' {
+  declare module.exports: $Exports<'eslint-plugin-jest/rules/__tests__/prefer-to-be-null.test'>;
+}
+declare module 'eslint-plugin-jest/rules/__tests__/prefer-to-be-undefined.test.js' {
+  declare module.exports: $Exports<'eslint-plugin-jest/rules/__tests__/prefer-to-be-undefined.test'>;
+}
+declare module 'eslint-plugin-jest/rules/__tests__/prefer-to-have-length.test.js' {
+  declare module.exports: $Exports<'eslint-plugin-jest/rules/__tests__/prefer-to-have-length.test'>;
+}
+declare module 'eslint-plugin-jest/rules/__tests__/valid-describe.test.js' {
+  declare module.exports: $Exports<'eslint-plugin-jest/rules/__tests__/valid-describe.test'>;
+}
+declare module 'eslint-plugin-jest/rules/__tests__/valid-expect-in-promise.test.js' {
+  declare module.exports: $Exports<'eslint-plugin-jest/rules/__tests__/valid-expect-in-promise.test'>;
+}
+declare module 'eslint-plugin-jest/rules/__tests__/valid-expect.test.js' {
+  declare module.exports: $Exports<'eslint-plugin-jest/rules/__tests__/valid-expect.test'>;
+}
+declare module 'eslint-plugin-jest/rules/consistent-test-it.js' {
+  declare module.exports: $Exports<'eslint-plugin-jest/rules/consistent-test-it'>;
+}
+declare module 'eslint-plugin-jest/rules/lowercase-name.js' {
+  declare module.exports: $Exports<'eslint-plugin-jest/rules/lowercase-name'>;
+}
+declare module 'eslint-plugin-jest/rules/no-disabled-tests.js' {
+  declare module.exports: $Exports<'eslint-plugin-jest/rules/no-disabled-tests'>;
+}
+declare module 'eslint-plugin-jest/rules/no-focused-tests.js' {
+  declare module.exports: $Exports<'eslint-plugin-jest/rules/no-focused-tests'>;
+}
+declare module 'eslint-plugin-jest/rules/no-hooks.js' {
+  declare module.exports: $Exports<'eslint-plugin-jest/rules/no-hooks'>;
+}
+declare module 'eslint-plugin-jest/rules/no-identical-title.js' {
+  declare module.exports: $Exports<'eslint-plugin-jest/rules/no-identical-title'>;
+}
+declare module 'eslint-plugin-jest/rules/no-jest-import.js' {
+  declare module.exports: $Exports<'eslint-plugin-jest/rules/no-jest-import'>;
+}
+declare module 'eslint-plugin-jest/rules/no-large-snapshots.js' {
+  declare module.exports: $Exports<'eslint-plugin-jest/rules/no-large-snapshots'>;
+}
+declare module 'eslint-plugin-jest/rules/no-test-prefixes.js' {
+  declare module.exports: $Exports<'eslint-plugin-jest/rules/no-test-prefixes'>;
+}
+declare module 'eslint-plugin-jest/rules/prefer-expect-assertions.js' {
+  declare module.exports: $Exports<'eslint-plugin-jest/rules/prefer-expect-assertions'>;
+}
+declare module 'eslint-plugin-jest/rules/prefer-to-be-null.js' {
+  declare module.exports: $Exports<'eslint-plugin-jest/rules/prefer-to-be-null'>;
+}
+declare module 'eslint-plugin-jest/rules/prefer-to-be-undefined.js' {
+  declare module.exports: $Exports<'eslint-plugin-jest/rules/prefer-to-be-undefined'>;
+}
+declare module 'eslint-plugin-jest/rules/prefer-to-have-length.js' {
+  declare module.exports: $Exports<'eslint-plugin-jest/rules/prefer-to-have-length'>;
+}
+declare module 'eslint-plugin-jest/rules/util.js' {
+  declare module.exports: $Exports<'eslint-plugin-jest/rules/util'>;
+}
+declare module 'eslint-plugin-jest/rules/valid-describe.js' {
+  declare module.exports: $Exports<'eslint-plugin-jest/rules/valid-describe'>;
+}
+declare module 'eslint-plugin-jest/rules/valid-expect-in-promise.js' {
+  declare module.exports: $Exports<'eslint-plugin-jest/rules/valid-expect-in-promise'>;
+}
+declare module 'eslint-plugin-jest/rules/valid-expect.js' {
+  declare module.exports: $Exports<'eslint-plugin-jest/rules/valid-expect'>;
 }
