@@ -1,5 +1,5 @@
-// flow-typed signature: 09ffb93507d568e073b5988fbf20dcf1
-// flow-typed version: <<STUB>>/eslint-plugin-jsx-a11y_v6.0.2/flow_v0.59.0
+// flow-typed signature: d23585c0713e9a1f229b346337433ac5
+// flow-typed version: <<STUB>>/eslint-plugin-jsx-a11y_v^6.0.3/flow_v0.68.0
 
 /**
  * This is an autogenerated libdef stub for:
@@ -199,6 +199,18 @@ declare module 'eslint-plugin-jsx-a11y/__tests__/src/util/getTabIndex-test' {
 }
 
 declare module 'eslint-plugin-jsx-a11y/__tests__/src/util/hasAccessibleChild-test' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/util/implicitRoles/input-test' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/util/implicitRoles/menu-test' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/util/implicitRoles/menuitem-test' {
   declare module.exports: any;
 }
 
@@ -725,6 +737,15 @@ declare module 'eslint-plugin-jsx-a11y/__tests__/src/util/getTabIndex-test.js' {
 }
 declare module 'eslint-plugin-jsx-a11y/__tests__/src/util/hasAccessibleChild-test.js' {
   declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/util/hasAccessibleChild-test'>;
+}
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/util/implicitRoles/input-test.js' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/util/implicitRoles/input-test'>;
+}
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/util/implicitRoles/menu-test.js' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/util/implicitRoles/menu-test'>;
+}
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/util/implicitRoles/menuitem-test.js' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/util/implicitRoles/menuitem-test'>;
 }
 declare module 'eslint-plugin-jsx-a11y/__tests__/src/util/isAbstractRole-test.js' {
   declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/util/isAbstractRole-test'>;
