@@ -1,5 +1,5 @@
-// flow-typed signature: ceee3246302833a712f4c145cab64258
-// flow-typed version: <<STUB>>/eslint_v4.11.0/flow_v0.59.0
+// flow-typed signature: fa3cca1017d7a5fcb5d3e569945693c3
+// flow-typed version: <<STUB>>/eslint_v^4.15.0/flow_v0.68.0
 
 /**
  * This is an autogenerated libdef stub for:
@@ -31,10 +31,6 @@ declare module 'eslint/conf/config-schema' {
 }
 
 declare module 'eslint/conf/default-cli-options' {
-  declare module.exports: any;
-}
-
-declare module 'eslint/conf/default-config-options' {
   declare module.exports: any;
 }
 
@@ -383,6 +379,10 @@ declare module 'eslint/lib/rules/id-length' {
 }
 
 declare module 'eslint/lib/rules/id-match' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/rules/implicit-arrow-linebreak' {
   declare module.exports: any;
 }
 
@@ -1322,6 +1322,10 @@ declare module 'eslint/lib/util/hash' {
   declare module.exports: any;
 }
 
+declare module 'eslint/lib/util/interpolate' {
+  declare module.exports: any;
+}
+
 declare module 'eslint/lib/util/keywords' {
   declare module.exports: any;
 }
@@ -1387,9 +1391,6 @@ declare module 'eslint/conf/config-schema.js' {
 }
 declare module 'eslint/conf/default-cli-options.js' {
   declare module.exports: $Exports<'eslint/conf/default-cli-options'>;
-}
-declare module 'eslint/conf/default-config-options.js' {
-  declare module.exports: $Exports<'eslint/conf/default-config-options'>;
 }
 declare module 'eslint/conf/environments.js' {
   declare module.exports: $Exports<'eslint/conf/environments'>;
@@ -1651,6 +1652,9 @@ declare module 'eslint/lib/rules/id-length.js' {
 }
 declare module 'eslint/lib/rules/id-match.js' {
   declare module.exports: $Exports<'eslint/lib/rules/id-match'>;
+}
+declare module 'eslint/lib/rules/implicit-arrow-linebreak.js' {
+  declare module.exports: $Exports<'eslint/lib/rules/implicit-arrow-linebreak'>;
 }
 declare module 'eslint/lib/rules/indent-legacy.js' {
   declare module.exports: $Exports<'eslint/lib/rules/indent-legacy'>;
@@ -2353,6 +2357,9 @@ declare module 'eslint/lib/util/glob.js' {
 }
 declare module 'eslint/lib/util/hash.js' {
   declare module.exports: $Exports<'eslint/lib/util/hash'>;
+}
+declare module 'eslint/lib/util/interpolate.js' {
+  declare module.exports: $Exports<'eslint/lib/util/interpolate'>;
 }
 declare module 'eslint/lib/util/keywords.js' {
   declare module.exports: $Exports<'eslint/lib/util/keywords'>;
