@@ -1,5 +1,5 @@
-// flow-typed signature: d82f0b441e4bfa97dde671f0d08bb191
-// flow-typed version: <<STUB>>/redux-thunk_v2.2.0/flow_v0.59.0
+// flow-typed signature: 4109af0de2abd079b0f6c3408cab1733
+// flow-typed version: <<STUB>>/redux-thunk_v^2.2.0/flow_v0.68.0
 
 /**
  * This is an autogenerated libdef stub for:
