@@ -11,21 +11,23 @@ import { sizeSnapshot } from 'rollup-plugin-size-snapshot';
 const pkg = require('./package.json');
 
 const input = './src/index.js';
-
 const extensions = ['.js', '.jsx'];
 
 // Treat as externals all not relative and not absolute paths
 // e.g. 'react'
-const isExternal = id => !id.startsWith('.') && !id.startsWith('/');
+const excludeAllExternals = id => !id.startsWith('.') && !id.startsWith('/');
 
 const getBabelOptions = () => ({
   exclude: 'node_modules/**',
   runtimeHelpers: true,
 });
 
-const checkSnapshot = process.env.SNAPSHOT === 'check';
+const shouldCheckSnapshot = process.env.SNAPSHOT === 'check';
 
 export default [
+  // Universal module definition (UMD) build
+  // - including console.* statements
+  // - conditionally used to check snapshot size
   {
     input,
     output: {
@@ -34,16 +36,17 @@ export default [
       name: 'ReactBeautifulDnd',
       globals: { react: 'React' },
     },
+    // Only deep dependency required is React
     external: ['react'],
     plugins: [
       babel(getBabelOptions()),
       resolve({ extensions }),
       commonjs({ include: 'node_modules/**' }),
       replace({ 'process.env.NODE_ENV': JSON.stringify('development') }),
-      sizeSnapshot({ updateSnapshot: !checkSnapshot }),
+      sizeSnapshot({ updateSnapshot: !shouldCheckSnapshot }),
     ],
   },
-
+  // Minified UMD build
   {
     input,
     output: {
@@ -52,6 +55,7 @@ export default [
       name: 'ReactBeautifulDnd',
       globals: { react: 'React' },
     },
+    // Only deep dependency required is React
     external: ['react'],
     plugins: [
       babel(getBabelOptions()),
@@ -59,29 +63,33 @@ export default [
       commonjs({ include: 'node_modules/**' }),
       replace({ 'process.env.NODE_ENV': JSON.stringify('production') }),
       strip({ debugger: true }),
-      sizeSnapshot({ updateSnapshot: !checkSnapshot }),
+      sizeSnapshot({ updateSnapshot: !shouldCheckSnapshot }),
       uglify(),
     ],
   },
-
+  // CommonJS (cjs) build
+  // - Keeping console.log statements
+  // - All external packages are not bundled
   {
     input,
     output: { file: pkg.main, format: 'cjs' },
-    external: isExternal,
+    external: excludeAllExternals,
     plugins: [
       resolve({ extensions }),
       babel(getBabelOptions()),
     ],
   },
-
+  // EcmaScript Module (esm) build
+  // - Keeping console.log statements
+  // - All external packages are not bundled
   {
     input,
     output: { file: pkg.module, format: 'es' },
-    external: isExternal,
+    external: excludeAllExternals,
     plugins: [
       resolve({ extensions }),
       babel(getBabelOptions()),
-      sizeSnapshot({ updateSnapshot: !checkSnapshot }),
+      sizeSnapshot({ updateSnapshot: !shouldCheckSnapshot }),
     ],
   },
 ];
