@@ -10,6 +10,9 @@ const flag: string = '__react-beautiful-dnd-debug-timings-hook__';
 
 const isTimingsEnabled = (): boolean => Boolean(window[flag]);
 
+// TEMP
+// window[flag] = true;
+
 export const start = (key: string) => {
   if (!isTimingsEnabled()) {
     return;
@@ -18,6 +21,11 @@ export const start = (key: string) => {
 
   records[key] = now;
 };
+
+type Style = {|
+  textColor: string,
+  symbol: string,
+|}
 
 export const finish = (key: string) => {
   if (!isTimingsEnabled()) {
@@ -33,8 +41,37 @@ export const finish = (key: string) => {
   }
 
   const result: number = now - previous;
+  const rounded: string = result.toFixed(2);
+
+  const style: Style = (() => {
+    if (result < 16) {
+      return {
+        textColor: 'green',
+        symbol: '✅',
+      };
+    }
+    if (result < 40) {
+      return {
+        textColor: 'orange',
+        symbol: '⚠️',
+      };
+    }
+    return {
+      textColor: 'red',
+      symbol: '❌',
+    };
+  })();
 
   // eslint-disable-next-line no-console
-  console.log(`%cTiming (${key}): %c${result} %cms`, 'font-weight: bold;', 'color: blue;', 'color: grey;');
+  console.log(`${style.symbol} %cTiming %c${rounded} %cms %c${key}`,
+    // title
+    'color: blue; font-weight: bold; ',
+    // result
+    `color: ${style.textColor}; font-size: 1.1em;`,
+    // ms
+    'color: grey;',
+    // key
+    'color: purple; font-weight: bold;',
+  );
 };
 
