@@ -1,14 +1,15 @@
 // @flow
 /* eslint-disable no-console */
+import * as timings from '../../debug/timings';
 import type { Action } from '../../types';
 
 export default () => (next: (Action) => mixed) => (action: Action): mixed => {
-  const key = `action: ${action.type}`;
-  console.time(key);
+  const key = `redux action: ${action.type}`;
+  timings.start(key);
 
   const result: mixed = next(action);
 
-  console.timeEnd(key);
+  timings.finish(key);
 
   return result;
 };
