@@ -2,6 +2,7 @@
 import type {
   DraggableId,
   DroppableId,
+  TypeId,
   DroppableDescriptor,
   DraggableDescriptor,
   DraggableDimension,
@@ -33,6 +34,7 @@ type State = {|
   isCollecting: boolean,
   scrollOptions: ?ScrollOptions,
   request: ?LiftRequest,
+  requestType: ?TypeId,
   frameId: ?AnimationFrameID,
 |}
 
@@ -48,6 +50,7 @@ export default (callbacks: Callbacks) => {
     isCollecting: false,
     scrollOptions: null,
     request: null,
+    requestType: null,
     frameId: null,
   };
 
@@ -62,13 +65,55 @@ export default (callbacks: Callbacks) => {
   const cancel = (...args: mixed[]) => {
     console.error(...args);
 
+    // We want to cancel the drag even if we are not collecting yet
+    // This is true when trying to lift something that has not been published
+    callbacks.cancel();
+
     if (!state.isCollecting) {
       return;
     }
 
     // eslint-disable-next-line no-use-before-define
     stopCollecting();
-    callbacks.cancel();
+  };
+
+  const cancelIfModifyingActiveDraggable = (descriptor: DraggableDescriptor) => {
+    if (!state.isCollecting) {
+      return;
+    }
+
+    const home: ?DroppableEntry = state.droppables[descriptor.droppableId];
+
+    // In React 16 children are mounted before parents are.
+    // This case can happen when a list of Draggables are being
+    // moved using a React.Portal.
+    if (!home) {
+      return;
+    }
+
+    // Adding something of a different type - not relevant to the drag
+    if (home.descriptor.type !== state.requestType) {
+      return;
+    }
+
+    // Technically we could let the drag go on - but this is being more explicit
+    // with consumers to prevent undesirable states
+    cancel('Adding or removing a Draggable during a drag is currently not supported');
+  };
+
+  const cancelIfModifyingActiveDroppable = (descriptor: DroppableDescriptor) => {
+    if (!state.isCollecting) {
+      return;
+    }
+
+    // Adding something of a different type - not relevant to the drag
+    // This can happen when dragging a Draggable that has a child Droppable
+    // when using a React.Portal as the child Droppable component will be remounted
+    if (descriptor.type !== state.requestType) {
+      return;
+    }
+
+    cancel('Adding or removing a Droppable during a drag is currently not supported');
   };
 
   const registerDraggable = (
@@ -77,14 +122,10 @@ export default (callbacks: Callbacks) => {
   ) => {
     const id: DraggableId = descriptor.id;
 
-    // Cannot register a draggable if no entry exists for the droppable
-    if (!state.droppables[descriptor.droppableId]) {
-      cancel(`Cannot register Draggable ${id} as there is no entry for the Droppable ${descriptor.droppableId}`);
-      return;
-    }
-
     // Not checking if the draggable already exists.
-    // This allows for overwriting in particular circumstances
+    // - This allows for overwriting in particular circumstances.
+    // Not checking if a parent droppable exists.
+    // - In React 16 children are mounted before their parents
 
     const entry: DraggableEntry = {
       descriptor,
@@ -99,11 +140,7 @@ export default (callbacks: Callbacks) => {
       draggables,
     });
 
-    if (!state.isCollecting) {
-      return;
-    }
-
-    console.warn('Adding a draggable during a drag is currently not supported');
+    cancelIfModifyingActiveDraggable(descriptor);
   };
 
   const registerDroppable = (
@@ -130,11 +167,7 @@ export default (callbacks: Callbacks) => {
       droppables,
     });
 
-    if (!state.isCollecting) {
-      return;
-    }
-
-    console.warn('Currently not supporting updating Droppables during a drag');
+    cancelIfModifyingActiveDroppable(descriptor);
   };
 
   const updateDroppableIsEnabled = (id: DroppableId, isEnabled: boolean) => {
@@ -146,6 +179,9 @@ export default (callbacks: Callbacks) => {
     if (!state.isCollecting) {
       return;
     }
+    // At this point a non primary droppable dimension might not yet be published
+    // but may have its enabled state changed. For now we still publish this change
+    // and let the reducer exit early if it cannot find the dimension in the state.
     callbacks.updateDroppableIsEnabled(id, isEnabled);
   };
 
@@ -198,11 +234,7 @@ export default (callbacks: Callbacks) => {
       draggables: newMap,
     });
 
-    if (!state.isCollecting) {
-      return;
-    }
-
-    console.warn('currently not supporting unmounting a Draggable during a drag');
+    cancelIfModifyingActiveDraggable(descriptor);
   };
 
   const unregisterDroppable = (descriptor: DroppableDescriptor) => {
@@ -232,11 +264,7 @@ export default (callbacks: Callbacks) => {
       droppables: newMap,
     });
 
-    if (!state.isCollecting) {
-      return;
-    }
-
-    console.warn('currently not supporting unmounting a Droppable during a drag');
+    cancelIfModifyingActiveDroppable(descriptor);
   };
 
   const getToBeCollected = (): UnknownDescriptorType[] => {
@@ -301,15 +329,9 @@ export default (callbacks: Callbacks) => {
       return;
     }
 
-    const draggableId: DraggableId = request.draggableId;
-
-    setState({
-      isCollecting: true,
-      request,
-    });
-
     const draggables: DraggableEntryMap = state.draggables;
     const droppables: DroppableEntryMap = state.droppables;
+    const draggableId: DraggableId = request.draggableId;
     const draggableEntry: ?DraggableEntry = draggables[draggableId];
 
     if (!draggableEntry) {
@@ -326,6 +348,12 @@ export default (callbacks: Callbacks) => {
       `);
       return;
     }
+
+    setState({
+      isCollecting: true,
+      request,
+      requestType: homeEntry.descriptor.type,
+    });
 
     // Get the minimum dimensions to start a drag
     const home: DroppableDimension = homeEntry.callbacks.getDimension();
