@@ -11,21 +11,20 @@ export default class Placeholder extends PureComponent<Props> {
     // We apply the margin separately to maintain margin collapsing
     // behavior of the original element
     const placeholder: PlaceholderType = this.props.placeholder;
-    const { top, left, bottom, right } = placeholder.margin;
-    const { width, height } = placeholder.paddingBox;
+    const { paddingBox, margin, display, tagName } = placeholder;
 
     const style = {
-      width,
-      height,
-      marginTop: top,
-      marginLeft: left,
-      marginBottom: bottom,
-      marginRight: right,
+      width: paddingBox.width,
+      height: paddingBox.height,
+      marginTop: margin.top,
+      marginLeft: margin.left,
+      marginBottom: margin.bottom,
+      marginRight: margin.right,
       pointerEvents: 'none',
       boxSizing: 'border-box',
+      display,
     };
-    return (
-      <div style={style} />
-    );
+
+    return React.createElement(tagName, { style });
   }
 }
