@@ -1,6 +1,5 @@
 // @flow
-/* eslint-disable react/no-multi-comp */
-import React, { Component } from 'react';
+import React from 'react';
 import styled from 'styled-components';
 import { Droppable, Draggable } from '../../../src';
 import QuoteItem from '../primatives/quote-item';
@@ -60,7 +59,7 @@ type QuoteListProps = {|
   autoFocusQuoteId: ?string,
 |}
 
-class InnerQuoteList extends Component<QuoteListProps> {
+class InnerQuoteList extends React.Component<QuoteListProps> {
   shouldComponentUpdate(nextProps: QuoteListProps) {
     if (nextProps.quotes !== this.props.quotes) {
       return true;
@@ -70,26 +69,19 @@ class InnerQuoteList extends Component<QuoteListProps> {
   }
 
   render() {
-    return (
-      <div>
-        {this.props.quotes.map((quote: Quote, index: number) => (
-          <Draggable key={quote.id} draggableId={quote.id} index={index}>
-            {(dragProvided: DraggableProvided, dragSnapshot: DraggableStateSnapshot) => (
-              <div>
-                <QuoteItem
-                  key={quote.id}
-                  quote={quote}
-                  isDragging={dragSnapshot.isDragging}
-                  provided={dragProvided}
-                  autoFocus={this.props.autoFocusQuoteId === quote.id}
-                />
-                {dragProvided.placeholder}
-              </div>
-          )}
-          </Draggable>
-        ))}
-      </div>
-    );
+    return this.props.quotes.map((quote: Quote, index: number) => (
+      <Draggable key={quote.id} draggableId={quote.id} index={index}>
+        {(dragProvided: DraggableProvided, dragSnapshot: DraggableStateSnapshot) => (
+          <QuoteItem
+            key={quote.id}
+            quote={quote}
+            isDragging={dragSnapshot.isDragging}
+            provided={dragProvided}
+            autoFocus={this.props.autoFocusQuoteId === quote.id}
+          />
+        )}
+      </Draggable>
+    ));
   }
 }
 
@@ -100,7 +92,7 @@ type InnerListProps = {|
   autoFocusQuoteId: ?string,
 |}
 
-class InnerList extends Component<InnerListProps> {
+class InnerList extends React.Component<InnerListProps> {
   render() {
     const { quotes, dropProvided, autoFocusQuoteId } = this.props;
     const title = this.props.title ? (
@@ -122,7 +114,7 @@ class InnerList extends Component<InnerListProps> {
   }
 }
 
-export default class QuoteList extends Component<Props> {
+export default class QuoteList extends React.Component<Props> {
   render() {
     const {
       ignoreContainerClipping,
