@@ -26,7 +26,7 @@ if (typeof document !== 'undefined' && typeof window !== 'undefined') {
 // setting up global enzyme
 const Enzyme = require('enzyme');
 
-const Adapter = require('enzyme-adapter-react-15');
+const Adapter = require('enzyme-adapter-react-16');
 
 Enzyme.configure({ adapter: new Adapter() });
 
