@@ -437,8 +437,7 @@ describe('Connected Droppable', () => {
     it('should render the child function when the parent renders', () => {
       const wrapper = mount(<App currentUser="Jake" />, contextOptions);
 
-      // initial render causes two renders due to setting child ref
-      expect(Person.prototype.render).toHaveBeenCalledTimes(2);
+      expect(Person.prototype.render).toHaveBeenCalledTimes(1);
       expect(wrapper.find(Person).props().name).toBe('Jake');
 
       wrapper.unmount();
@@ -449,8 +448,7 @@ describe('Connected Droppable', () => {
 
       forceUpdate(wrapper);
 
-      // initial render causes two renders due to setting child ref
-      expect(Person.prototype.render).toHaveBeenCalledTimes(3);
+      expect(Person.prototype.render).toHaveBeenCalledTimes(2);
       expect(wrapper.find(Person).props().name).toBe('Jake');
 
       wrapper.unmount();
@@ -463,8 +461,7 @@ describe('Connected Droppable', () => {
         currentUser: 'Finn',
       });
 
-      // initial render causes two renders due to setting child ref
-      expect(Person.prototype.render).toHaveBeenCalledTimes(3);
+      expect(Person.prototype.render).toHaveBeenCalledTimes(2);
       expect(wrapper.find(Person).props().name).toBe('Finn');
 
       wrapper.unmount();
